Refactoring Timer.

+++ basic_types.h
@@ -12,13 +12,13 @@ typedef unsigned short ushort;
 typedef unsigned int uint;
 typedef unsigned long ulong;
 
-typedef long Pos;
-typedef short PosValue;
-typedef short Player;
+typedef int Pos;
+typedef int PosValue;
+typedef int Player;
 typedef int Score;
-typedef short Coord;
+typedef int Coord;
 typedef size_t Hash;
-typedef size_t Time;
+typedef int Time;
 
 #define SCORE_INFINITY numeric_limits<Score>::max()
